requirements.txt
future
bilby>=2.4.1
bilby_pipe==1.5.0
colorcet
numpy>2.0
matplotlib>3.8
scipy>=1.10
pandas>2.0
astropy>=5.3.4
pymultinest
sncosmo
dust_extinction
arviz
p_tqdm
tornado
notebook
ligo.skymap
healpy
dustmaps
